DocExtensions

Documentation for DocExtensions.

DocExtensions.DocExtensions

DocExtensions.expand_reflinks

DocExtensions.DocExtensions — Module

Module DocExtensions

Provides extensions to documentation tools as Documenter.jl and Literate.jl, including a simple file processor for pre- or post-processing of files.

DocExtensions.expand_reflinks — Method

expand_reflinks(reflinks; filepattern=".MD", outfilepattern=".md", rootdir=".")

Expand the reference links reflinks in the file or collection of files in directory rootdir defined by filepattern, writing the result back to the file(s) specified with outfilepattern; for both filepattern and outfilepattern currently supported is a file name extension (e.g. filepattern = ".MD" and outfilepattern = ".md") or simply an explicit file name.

Arguments

reflinks::Dict{<:AbstractString, <:AbstractString}: the reference links to be expanded.

Keyword Arguments

`filepattern::AbstractString=".MD": a file extension or file name defining the files to be processed.

`outfilepattern::AbstractString=".MD": a file extension or file name defining the name(s) of the output files.

rootdir::AbstractString=".": the root directory of the files to be processed.